chore: split RootOutput type
