+++ internal/generator/types.go
@@ -1,0 +1,10 @@
+package generator
+
+type RootOutput struct {
+	OutputBasedir       string
+	ImportPackages      map[string]string
+	Providers           []string
+	InjectedServices    []string
+	InjectedControllers []string
+	RoutesCode          []string
+}
